使用PropertyUtils替换BeanUntils以解决非String属性的Unique校验问题，并增加测试案例测试非String属性的Unique校验
