--- src/test/java/org/ibankapp/base/persistence/test/TestModel.java
@@ -1,35 +1,0 @@
-/*
- * iBankApp
- *
- * License : Apache License,Version 2.0, January 2004
- *
- * See the LICENSE file in English or LICENSE.zh_CN in chinese
- * in the root directory or <http://www.apache.org/licenses/>.
- */
-
-package org.ibankapp.base.persistence.test;
-
-import org.ibankapp.base.persistence.Model;
-import org.ibankapp.base.validation.constraints.Unique;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "TEST_MODEL")
-@Unique(properties = {"name"},message = "姓名重复")
-public class TestModel extends Model {
-
-    private String name;
-
-    @Column
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-}
