+++ src/test/java/org/ibankapp/base/persistence/test/model/TestModel.java
@@ -7,7 +7,16 @@
  * in the root directory or <http://www.apache.org/licenses/>.
  */
 
-package org.ibankapp.base.persistence.test;
+/*
+ * iBankApp
+ *
+ * License : Apache License,Version 2.0, January 2004
+ *
+ * See the LICENSE file in English or LICENSE.zh_CN in chinese
+ * in the root directory or <http://www.apache.org/licenses/>.
+ */
+
+package org.ibankapp.base.persistence.test.model;
 
 import org.ibankapp.base.persistence.Model;
 import org.ibankapp.base.validation.constraints.Unique;
@@ -18,8 +27,8 @@ import javax.persistence.Table;
 
 @Entity
 @Table(name = "TEST_MODEL")
-@Unique(properties = {"type"},message = "姓名重复")
-public class TestModelWithNoExistUniqueColumn extends Model {
+@Unique(properties = {"name"},message = "姓名重复")
+public class TestModel extends Model {
 
     private String name;
