+++ src/test/java/org/ibankapp/base/validation/test/model/TestModelWithInheritedUnique.java
@@ -7,7 +7,16 @@
  * in the root directory or <http://www.apache.org/licenses/>.
  */
 
-package org.ibankapp.base.persistence.test;
+/*
+ * iBankApp
+ *
+ * License : Apache License,Version 2.0, January 2004
+ *
+ * See the LICENSE file in English or LICENSE.zh_CN in chinese
+ * in the root directory or <http://www.apache.org/licenses/>.
+ */
+
+package org.ibankapp.base.validation.test.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
