+++ src/test/java/org/ibankapp/base/validation/test/model/TestModelWithNoExistUniqueColumn.java
@@ -7,26 +7,30 @@
  * in the root directory or <http://www.apache.org/licenses/>.
  */
 
+/*
+ * iBankApp
+ *
+ * License : Apache License,Version 2.0, January 2004
+ *
+ * See the LICENSE file in English or LICENSE.zh_CN in chinese
+ * in the root directory or <http://www.apache.org/licenses/>.
+ */
 
-package org.ibankapp.base.persistence.test;
+package org.ibankapp.base.validation.test.model;
 
 import org.ibankapp.base.persistence.Model;
 import org.ibankapp.base.validation.constraints.Unique;
-import org.ibankapp.base.validation.constraints.Uniques;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "TEST_MODELTESTWITHUNIQUES")
-@Uniques(constraints = {@Unique(properties = {"name"}, message = "姓名重复"),
-        @Unique(properties = "type", message = "类型重复")})
-public class TestModelWithUniques extends Model {
+@Table(name = "TEST_MODEL")
+@Unique(properties = {"type"},message = "姓名重复")
+public class TestModelWithNoExistUniqueColumn extends Model {
 
     private String name;
-
-    private String type;
 
     @Column
     public String getName() {
@@ -37,12 +41,4 @@ public class TestModelWithUniques extends Model {
         this.name = name;
     }
 
-    @Column
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
 }
